map: fixed origin set with python3

+++ MAVProxy/modules/mavproxy_map/__init__.py
@@ -573,9 +573,9 @@ class MapModule(mp_module.MPModule):
         print("Setting origin to: ", lat, lon, alt)
         self.master.mav.set_gps_global_origin_send(
             self.settings.target_system,
-            lat*10000000, # lat
-            lon*10000000, # lon
-            alt*1000) # param7
+            int(lat*10000000), # lat
+            int(lon*10000000), # lon
+            int(alt*1000)) # param7
 
     def cmd_set_originpos(self, args):
         '''called when user selects "Set Origin" on map'''
@@ -583,8 +583,8 @@ class MapModule(mp_module.MPModule):
         print("Setting origin to: ", lat, lon)
         self.master.mav.set_gps_global_origin_send(
             self.settings.target_system,
-            lat*10000000, # lat
-            lon*10000000, # lon
+            int(lat*10000000), # lat
+            int(lon*10000000), # lon
             0*1000) # no height change
 
     def cmd_zoom(self, args):
